requirements.txt
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.7
idna==3.3
python-dotenv==0.19.2
requests==2.26.0
secure-smtplib==0.1.1
soupsieve==2.3.1
urllib3==1.26.7
